refactor(controller): move /RaceController.kt to /controller/RaceController.kt
