+++ src/main/kotlin/racingcar/controller/RaceController.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.controller
 
 import camp.nextstep.edu.missionutils.Randoms
 import racingcar.domain.Car
